used responseEntity in this project and added dependency of lombok

+++ src/main/java/com/edigest/journalApp/controller/JournalEntryController.java
@@ -1,21 +1,25 @@
 package com.edigest.journalApp.controller;
 
+import ch.qos.logback.core.pattern.parser.OptionTokenizer;
 import com.edigest.journalApp.entity.JournalEntry;
 import com.edigest.journalApp.service.JournalEntryService;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.graphql.GraphQlProperties;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/journal")
-public class JournalEntryController {
+public class JournalEntryController{
 
     @Autowired
     private JournalEntryService journalEntryService;
-
 
     @GetMapping("/getAll")
     public List<JournalEntry> getAllEntries(){
@@ -30,17 +34,26 @@ public class JournalEntryController {
     }
 
     @GetMapping("/id/{id}")
-    public JournalEntry getJournalEntryById(@PathVariable ObjectId id){
-        return journalEntryService.findById(id).orElse(null);
+    public ResponseEntity<JournalEntry> getJournalEntryById(@PathVariable ObjectId id){
+        Optional<JournalEntry> journalEntry = journalEntryService.findById(id);
+        return journalEntry.isPresent()
+                ? new ResponseEntity<>(journalEntry.get() , HttpStatus.OK)
+                : new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping("/id/{id}")
-    public String deleteById(@PathVariable ObjectId id) {
-        journalEntryService.deleteById(id);
-        return ("Entry with " + id + " deleted successfully!");
+    public ResponseEntity<String> deleteById(@PathVariable ObjectId id) {
+        Optional<JournalEntry> journalEntry = journalEntryService.findById(id);
+
+        if(journalEntry.isPresent()){
+            journalEntryService.deleteById(id);
+            return new ResponseEntity<> ("Entry with " + id + " deleted successfully!", HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>("Entry with " + id + " not found!" , HttpStatus.NOT_FOUND);
+        }
     }
 
-    @PutMapping("id/{id}")
+    @PutMapping("/id/{id}")
     public String updateById(@PathVariable ObjectId id , @RequestBody JournalEntry newEntry){
         JournalEntry old = journalEntryService.findById(id).orElse(null);
         if(old!=null){
